requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-braces==1.14.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytest-pylint==0.15.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
zipp==3.0.0
